Poprawki

+++ 35 Prostopadłe/main.js
@@ -6,10 +6,10 @@ const wynik = document.querySelector('#wynik');
 const btn = document.querySelector('button');
 
 btn.addEventListener('click', function () {
-    let a1 = parseInt(row_a1.value);
-    let a2 = parseInt(row_a2.value);
-    let b1=parseInt(row_b1.value);
-    let b2=parseInt(row_b2.value);
+    let a1 = parseFloat(row_a1.value);
+    let a2 = parseFloat(row_a2.value);
+    let b1=parseFloat(row_b1.value);
+    let b2=parseFloat(row_b2.value);
     if(a1*a2===-1){
         wynik.innerHTML = 'Dla równań y='+a1+ 'x+' +b1+ ' oraz y=' +a2+ 'x+' +b2+' proste są prostopadłe'
     } else {
